cleanup() rename files to mtach new products collection
