+++ src/modules/root.js
@@ -2,9 +2,9 @@ import { combineEpics } from 'redux-observable';
 import reducer from '../reducers';
 import interestEpic from '../epics/interestEpic';
 import transactionLogEpic from '../epics/transaction';
-import getAccountsEpic from '../epics/getAccountsEpic';
+import getProductsEpic from '../epics/getProductsEpic';
 import getAvailableProductsEpic from '../epics/getAvailableProductsEpic';
-import getAccountEpic from '../epics/getAccountEpic';
+import getProductEpic from '../epics/getProductEpic';
 import logOutEpic from '../epics/logOutEpic';
 import logInEpic from '../epics/logInEpic';
 import signUpEpic from '../epics/signUpEpic';
@@ -24,7 +24,7 @@ import { catchError } from 'rxjs/operators';
  */
 export const rootEpic = (action$, store$, dependencies) =>
   combineEpics(
-    getAccountsEpic,
+    getProductsEpic,
     getAvailableProductsEpic,
     addProductEpic,
     interestEpic,
@@ -34,7 +34,7 @@ export const rootEpic = (action$, store$, dependencies) =>
     signUpEpic,
     signedUpEpic,
     //getUserDataEpic,
-    getAccountEpic,
+    getProductEpic,
   )(action$, store$, dependencies).pipe(
     catchError((error, source) => {
       console.error(error);
